+++ src/constants.ts
@@ -11,6 +11,7 @@ export const projectRuntimeSetting: string = 'projectRuntime';
 export const templateFilterSetting: string = 'templateFilter';
 export const deploySubpathSetting: string = 'deploySubpath';
 export const templateVersionSetting: string = 'templateVersion';
+export const preDeployTaskSetting: string = 'preDeployTask';
 export const v1ReleaseVersion: string = '1.0.12'; // known stable version
 export const betaReleaseVersion: string = '2.0.1-beta.25'; // known stable version
 
@@ -65,3 +66,6 @@ export enum ScmType {
     LocalGit = 'LocalGit',
     GitHub = 'GitHub'
 }
+
+export const publishTaskId: string = 'publish';
+export const installExtensionsId: string = 'installExtensions';